change in auto_install.sh

+++ auto_install.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 echo "Auto install of all my need tool for a server"
-# echo "ad augusta per angusta"
+echo "This script is for debian based distribution"; sleep 1
 if [[ $EUID -ne 0 ]]; then
   echo "You must be a root user" 2>&1
   exit 1
@@ -20,13 +20,33 @@ wget https://raw.githubusercontent.com/rogafe/server_dotefile/master/zshrc
 mv zshrc ~/.zshrc
 echo "install of antigen done"
 cd /tmp
-echo "[installing docker]"
-curl -fsSL get.docker.com -o get-docker.sh && sh get-docker.sh
-echo "[docker installed]"
-sleep 2
-echo "[installing spf13-vim]"
-curl http://j.mp/spf13-vim3 -L -o - | sh
-echo "[all done closing the ssh shell]"
-sleep 2
+
+echo "[?] Docker ?"
+
+read -p "Install docker ? (y/n)" answer
+case ${answer:0:1} in
+    y|Y )
+        echo "[installing docker]"
+        curl -fsSL get.docker.com -o get-docker.sh && sh get-docker.sh
+        echo "[docker installed ]" 
+    * )
+        echo No
+        ;;
+esac
+       sleep 2
+
+echo "[?] spf13-vim "
+read -p "Install spf13-vim ? (y/n)"
+case ${answer:0:1} in
+    y|Y )
+        echo "[installing spf13-vim]"
+        curl http://j.mp/spf13-vim3 -L -o - | sh
+        echo "[all done closing the ssh shell]"
+        echo "Some time shell not work after this so I  close the remote shell"
+        sleep 2
+        exit
+        * )
+esac
 exit
+
 fi
